requirements/ci.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements/ci.txt requirements/base.txt requirements/test-tools.in
#
amqp==5.0.3               # via -r requirements/base.txt, kombu
beautifulsoup4==4.9.0     # via webtest
billiard==3.6.3.0         # via -r requirements/base.txt, celery
celery==5.0.5             # via -r requirements/base.txt
certifi==2020.4.5.1       # via -r requirements/base.txt, elastic-apm, requests, sentry-sdk
chardet==3.0.4            # via -r requirements/base.txt, requests
click-didyoumean==0.0.3   # via -r requirements/base.txt, celery
click-plugins==1.1.1      # via -r requirements/base.txt, celery
click-repl==0.1.6         # via -r requirements/base.txt, celery
click==7.1.2              # via -r requirements/base.txt, celery, click-didyoumean, click-plugins, click-repl
coverage==4.5.4           # via -r requirements/test-tools.in
cssselect==1.1.0          # via pyquery
django-admin-index==1.4.0  # via -r requirements/base.txt
django-appconf==1.0.4     # via -r requirements/base.txt, django-axes
django-axes==5.3.1        # via -r requirements/base.txt
django-choices==1.7.1     # via -r requirements/base.txt, zgw-consumers
django-compat==1.0.15     # via -r requirements/base.txt, django-hijack, django-hijack-admin
django-hijack-admin==2.1.10  # via -r requirements/base.txt
django-hijack==2.1.10     # via -r requirements/base.txt, django-hijack-admin
django-ipware==2.1.0      # via -r requirements/base.txt, django-axes
django-jenkins==0.110.0   # via -r requirements/test-tools.in
django-ordered-model==3.3.0  # via -r requirements/base.txt, django-admin-index
django-redis==4.11.0      # via -r requirements/base.txt
django-relativedelta==1.1.2  # via -r requirements/base.txt, zgw-consumers
django-rosetta==0.9.4     # via -r requirements/base.txt
django-sendfile==0.3.11   # via -r requirements/base.txt
django-sniplates==0.7.0   # via -r requirements/base.txt
django-solo==1.1.3        # via -r requirements/base.txt, zgw-consumers
django-webtest==1.9.7     # via -r requirements/test-tools.in
django==2.2.12            # via -r requirements/base.txt, django-appconf, django-axes, django-choices, django-jenkins, django-redis, django-relativedelta, django-rosetta, django-sendfile, django-sniplates, zgw-consumers
elastic-apm==5.5.2        # via -r requirements/base.txt
factory-boy==2.12.0       # via -r requirements/test-tools.in
faker==4.0.3              # via factory-boy
freezegun==0.3.15         # via -r requirements/test-tools.in
gemma-zds-client==0.13.3  # via -r requirements/base.txt, zgw-consumers
idna==2.9                 # via -r requirements/base.txt, requests
importlib-metadata==3.4.0  # via -r requirements/base.txt, kombu
kombu==5.0.2              # via -r requirements/base.txt, celery
lxml==4.6.2               # via -r requirements/base.txt, pyquery
pillow==7.1.1             # via -r requirements/base.txt
polib==1.1.0              # via -r requirements/base.txt, django-rosetta
prompt-toolkit==3.0.11    # via -r requirements/base.txt, click-repl
psycopg2==2.8.5           # via -r requirements/base.txt
pyjwt==1.7.1              # via -r requirements/base.txt, gemma-zds-client
pyquery==1.4.1            # via -r requirements/test-tools.in
python-dateutil==2.8.1    # via -r requirements/base.txt, django-relativedelta, faker, freezegun
python-decouple==3.3      # via -r requirements/base.txt
python-dotenv==0.12.0     # via -r requirements/base.txt
pytz==2019.3              # via -r requirements/base.txt, celery, django
pyyaml==5.3.1             # via -r requirements/base.txt, gemma-zds-client
redis==3.4.1              # via -r requirements/base.txt, django-redis
requests-mock==1.7.0      # via -r requirements/test-tools.in
requests==2.23.0          # via -r requirements/base.txt, django-rosetta, gemma-zds-client, requests-mock, zgw-consumers
sentry-sdk==0.14.3        # via -r requirements/base.txt
six==1.14.0               # via -r requirements/base.txt, click-repl, django-choices, django-compat, django-rosetta, freezegun, python-dateutil, requests-mock, webtest
soupsieve==2.0            # via beautifulsoup4
sqlparse==0.3.1           # via -r requirements/base.txt, django
tblib==1.6.0              # via -r requirements/test-tools.in
text-unidecode==1.3       # via faker
typing-extensions==3.7.4.3  # via -r requirements/base.txt, importlib-metadata
urllib3==1.25.8           # via -r requirements/base.txt, elastic-apm, requests, sentry-sdk
uwsgi==2.0.18             # via -r requirements/base.txt
vine==5.0.0               # via -r requirements/base.txt, amqp, celery
waitress==1.4.3           # via webtest
wcwidth==0.2.5            # via -r requirements/base.txt, prompt-toolkit
webob==1.8.6              # via webtest
webtest==2.0.34           # via django-webtest
zgw-consumers==0.13.1     # via -r requirements/base.txt
zipp==3.4.0               # via -r requirements/base.txt, importlib-metadata
